wikibase_integration/src/requirements.txt
wikibaseintegrator
pandas
pathlib
datetime
pandas
numpy
openai
tiktoken
spacy
nltk
word2number
tqdm
python-dotenv
psycopg2
mwparserfromhell>=0.5.0
pywikibot